+++ app/assets/javascripts/oxalis/view/action-bar/tracing_actions_view.js
@@ -13,6 +13,7 @@ import ButtonComponent from "oxalis/view/components/button_component";
 import messages from "messages";
 import api from "oxalis/api/internal_api";
 import { undoAction, redoAction } from "oxalis/model/actions/save_actions";
+import { setVersionRestoreVisibilityAction } from "oxalis/model/actions/ui_actions";
 import { copyAnnotationToUserAccount, finishAnnotation } from "admin/admin_rest_api";
 import { location } from "libs/window";
 import type { OxalisState, RestrictionsAndSettingsType, TaskType } from "oxalis/store";
@@ -52,6 +53,11 @@ class TracingActionsView extends PureComponent<StateProps, State> {
 
   handleUndo = () => {
     Store.dispatch(undoAction());
+  };
+
+  handleRestore = async () => {
+    await Model.save();
+    Store.dispatch(setVersionRestoreVisibilityAction(true));
   };
 
   handleRedo = () => {
@@ -231,6 +237,15 @@ class TracingActionsView extends PureComponent<StateProps, State> {
           isVisible={this.state.isMergeModalOpen}
           onOk={this.handleMergeClose}
         />,
+      );
+    }
+
+    if (isSkeletonMode && restrictions.allowUpdate) {
+      elements.push(
+        <Menu.Item key="restore-button" onClick={this.handleRestore}>
+          <Icon type="bars" theme="outlined" />
+          Restore Older Version
+        </Menu.Item>,
       );
     }